pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "seoul-metro-realtime"
version = "0.1.0"
description = "Skill scripts for Seoul Metro realtime arrivals"
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.27,<0.29",
  "python-dotenv>=1.0,<2.0",
]

[project.scripts]
seoul-metro-realtime = "seoul_metro_realtime.get_arrivals:main"

[dependency-groups]
dev = [
  "pytest>=8.0,<9.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/seoul_metro_realtime"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
